requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
python-dateutil>=2.8.2
descope>=1.0.0
PyJWT>=2.8.0
jinja2>=3.0.0